LibGUI+Spreadsheet: (WIP) Fix InfinitelyScrollableTableView shift select
This is a really really early and quite broken patch to fix shift
selections in Spreadsheet::InfinitelyScrollableTableView.

Currently the selection breaks randomly due to the cursor being set at
the wrong time. Fixing this is quite involved as AbstractView would have
to either be smarter about the Spreadsheet usecase or
InfinitelyScrollableTableView would have to handle a lot more selection
logic thats currently part of AbstractView itself.

The patch also includes a bunch of debugging statements